refactor: グループ作成・編集のcssを統一する #119
