+++ frontend/src/app/routes/app/management/groupEdit.tsx
@@ -12,7 +12,7 @@ import { useOutletContext, useNavigate, useLocation } from "react-router";
 import { useMembers } from "@/features/management/hooks/useMember";
 import { flattenGroups } from "@/features/management/utils/flattenGroups";
 import { findParentGroup } from "@/features/management/utils/findParentGroup";
-import styles from "@/features/management/style/groupNew.module.css";
+import styles from "@/features/management/style/groupForm.module.css";
 import { selectStyle } from "@/features/management/style/multiSelectStyle";
 
 const GroupEdit = () => {
